+++ app/screens/AppNavigator/screens/Balances/screens/ConvertConfirmationScreen.tsx
@@ -7,13 +7,11 @@ import React, { Dispatch, useEffect, useState } from 'react'
 import { ScrollView, View } from 'react-native'
 import { useDispatch, useSelector } from 'react-redux'
 import { Logging } from '../../../../../api'
-import {
-  ConfirmTitle,
-  NumberRow,
-  SubmitButtonGroup,
-  TokenBalanceRow
-} from '../../../../../components/ConfirmComponents'
+import { NumberRow } from '../../../../../components/NumberRow'
 import { SectionTitle } from '../../../../../components/SectionTitle'
+import { SubmitButtonGroup } from '../../../../../components/SubmitButtonGroup'
+import { SummaryTitle } from '../../../../../components/SummaryTitle'
+import { TokenBalanceRow } from '../../../../../components/TokenBalanceRow'
 import { RootState } from '../../../../../store'
 import { hasTxQueued as hasBroadcastQueued } from '../../../../../store/ocean'
 import { hasTxQueued, transactionQueue } from '../../../../../store/transaction_queue'
@@ -68,7 +66,7 @@ export function ConvertConfirmationScreen ({ route }: Props): JSX.Element {
 
   return (
     <ScrollView style={tailwind('bg-gray-100 pb-4')}>
-      <ConfirmTitle
+      <SummaryTitle
         title={translate('screens/ConvertConfirmationScreen', 'YOU ARE CONVERTING')}
         testID='text_convert_amount' amount={amount}
         suffix={` ${mode === 'utxosToAccount' ? 'DFI (UTXO)' : 'DFI (Token)'}`}
