+++ app/screens/AppNavigator/screens/Dex/DexConfirmAddLiquidity.tsx
@@ -10,8 +10,11 @@ import { ScrollView } from 'react-native'
 import { useDispatch, useSelector } from 'react-redux'
 import { Dispatch } from 'redux'
 import { Logging } from '../../../../api'
-import { ConfirmTitle, NumberRow, SubmitButtonGroup, TokenBalanceRow } from '../../../../components/ConfirmComponents'
+import { NumberRow } from '../../../../components/NumberRow'
 import { SectionTitle } from '../../../../components/SectionTitle'
+import { SubmitButtonGroup } from '../../../../components/SubmitButtonGroup'
+import { SummaryTitle } from '../../../../components/SummaryTitle'
+import { TokenBalanceRow } from '../../../../components/TokenBalanceRow'
 import { RootState } from '../../../../store'
 import { hasTxQueued as hasBroadcastQueued } from '../../../../store/ocean'
 import { hasTxQueued, transactionQueue } from '../../../../store/transaction_queue'
@@ -97,7 +100,7 @@ export function ConfirmAddLiquidityScreen (props: Props): JSX.Element {
 
   return (
     <ScrollView testID='confirm-root' style={tailwind('bg-gray-100 pb-4')}>
-      <ConfirmTitle
+      <SummaryTitle
         title={translate('screens/ConfirmAddLiq', 'YOU ARE ADDING')}
         testID='text_add_amount' amount={lmTokenAmount}
         suffix={` ${symbol}`}
@@ -135,7 +138,7 @@ export function ConfirmAddLiquidityScreen (props: Props): JSX.Element {
       />
       <NumberRow
         lhs={translate('screens/ConfirmAddLiq', 'Share of pool')}
-        rightHandElements={[{ value: percentage.times(100).toFixed(8), suffix: ' %', testID: 'percentage_pool' }]}
+        rightHandElements={[{ value: percentage.times(100).toFixed(8), suffix: '%', testID: 'percentage_pool' }]}
       />
       <NumberRow
         lhs={translate('screens/ConfirmAddLiq', `Pooled ${aSymbol}`)}
