+++ app/screens/AppNavigator/screens/Dex/DexConfirmRemoveLiquidity.tsx
@@ -8,8 +8,11 @@ import React, { useEffect, useState } from 'react'
 import { ScrollView } from 'react-native'
 import { useDispatch, useSelector } from 'react-redux'
 import { Dispatch } from 'redux'
-import { ConfirmTitle, NumberRow, SubmitButtonGroup, TokenBalanceRow } from '../../../../components/ConfirmComponents'
+import { NumberRow } from '../../../../components/NumberRow'
 import { SectionTitle } from '../../../../components/SectionTitle'
+import { SubmitButtonGroup } from '../../../../components/SubmitButtonGroup'
+import { SummaryTitle } from '../../../../components/SummaryTitle'
+import { TokenBalanceRow } from '../../../../components/TokenBalanceRow'
 import { RootState } from '../../../../store'
 import { hasTxQueued as hasBroadcastQueued } from '../../../../store/ocean'
 import { hasTxQueued, transactionQueue } from '../../../../store/transaction_queue'
@@ -69,7 +72,7 @@ export function RemoveLiquidityConfirmScreen ({ route }: Props): JSX.Element {
 
   return (
     <ScrollView style={tailwind('bg-gray-100 pb-4')}>
-      <ConfirmTitle
+      <SummaryTitle
         title={translate('screens/ConfirmRemoveLiquidity', 'YOU ARE REMOVING')}
         testID='text_remove_amount' amount={amount}
         suffix={` ${pair.symbol}`}
