+++ cypress/integration/functional/wallet/dex/add_liquidity.spec.ts
@@ -17,10 +17,8 @@ context('Wallet - DEX - Add Liquidity', () => {
     cy.getByTestID('MAX_amount_button').first().click()
     cy.getByTestID('token_input_primary').should('have.value', '10.00000000')
     cy.getByTestID('token_input_secondary').should('have.value', '10.00000000')
-    cy.getByTestID('a_per_b_price').contains('1.00000000')
-    cy.getByTestID('a_per_b_unit').contains('BTC per DFI')
-    cy.getByTestID('b_per_a_price').contains('1.00000000')
-    cy.getByTestID('b_per_a_unit').contains('DFI per BTC')
+    cy.getByTestID('a_per_b_price').contains('1.00000000 BTC per DFI')
+    cy.getByTestID('b_per_a_price').contains('1.00000000 DFI per BTC')
     cy.getByTestID('share_of_pool').contains('1.00000000%')
   })
 
@@ -29,10 +27,8 @@ context('Wallet - DEX - Add Liquidity', () => {
     cy.getByTestID('50%_amount_button').first().click()
     cy.getByTestID('token_input_primary').should('have.value', '5.00000000')
     cy.getByTestID('token_input_secondary').should('have.value', '5.00000000')
-    cy.getByTestID('a_per_b_price').contains('1.00000000')
-    cy.getByTestID('a_per_b_unit').contains('BTC per DFI')
-    cy.getByTestID('b_per_a_price').contains('1.00000000')
-    cy.getByTestID('b_per_a_unit').contains('DFI per BTC')
+    cy.getByTestID('a_per_b_price').contains('1.00000000 BTC per DFI')
+    cy.getByTestID('b_per_a_price').contains('1.00000000 DFI per BTC')
     cy.getByTestID('share_of_pool').contains('0.50000000%')
   })
 
@@ -43,10 +39,8 @@ context('Wallet - DEX - Add Liquidity', () => {
     cy.getByTestID('token_input_primary').clear().type('3')
     cy.getByTestID('token_input_secondary').should('have.value', '3.00000000')
 
-    cy.getByTestID('a_per_b_price').contains('1.00000000')
-    cy.getByTestID('a_per_b_unit').contains('BTC per DFI')
-    cy.getByTestID('b_per_a_price').contains('1.00000000')
-    cy.getByTestID('b_per_a_unit').contains('DFI per BTC')
+    cy.getByTestID('a_per_b_price').contains('1.00000000 BTC per DFI')
+    cy.getByTestID('b_per_a_price').contains('1.00000000 DFI per BTC')
     cy.getByTestID('share_of_pool').contains('0.30000000%')
   })
 
@@ -55,10 +49,8 @@ context('Wallet - DEX - Add Liquidity', () => {
 
     cy.getByTestID('token_input_primary').should('have.value', '2.00000000')
 
-    cy.getByTestID('a_per_b_price').contains('1.00000000')
-    cy.getByTestID('a_per_b_unit').contains('BTC per DFI')
-    cy.getByTestID('b_per_a_price').contains('1.00000000')
-    cy.getByTestID('b_per_a_unit').contains('DFI per BTC')
+    cy.getByTestID('a_per_b_price').contains('1.00000000 BTC per DFI')
+    cy.getByTestID('b_per_a_price').contains('1.00000000 DFI per BTC')
     cy.getByTestID('share_of_pool').contains('0.20000000%')
     cy.getByTestID('button_continue_add_liq').click()
   })
@@ -69,7 +61,7 @@ context('Wallet - DEX - Add Liquidity', () => {
     cy.getByTestID('a_amount').contains('2.00000000')
     cy.getByTestID('b_amount_unit').contains('BTC')
     cy.getByTestID('b_amount').contains('2.00000000')
-    cy.getByTestID('percentage_pool').contains('0.20000000 %')
+    cy.getByTestID('percentage_pool').contains('0.20000000%')
     cy.getByTestID('button_cancel_add').click()
   })
 
@@ -81,7 +73,7 @@ context('Wallet - DEX - Add Liquidity', () => {
     cy.getByTestID('a_amount').contains('10.00000000')
     cy.getByTestID('b_amount_unit').contains('BTC')
     cy.getByTestID('b_amount').contains('10.00000000')
-    cy.getByTestID('percentage_pool').contains('1.00000000 %')
+    cy.getByTestID('percentage_pool').contains('1.00000000%')
     cy.getByTestID('button_confirm_add').click().wait(3000)
     cy.closeOceanInterface()
   })
